+++ src/Model/ImageModelTrait.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Shw\Gallery\Model;
+
+use Pagekit\Application as App;
+use Pagekit\Database\ORM\ModelTrait;
+
+trait ImageModelTrait
+{
+    use ModelTrait;
+
+    /**
+     * @Saving
+     */
+    public static function saving($event, Image $image)
+    {
+        $image->modified = new \DateTime();
+    }
+}